packages/backend/pyproject.toml
[project]
name = "contextforge-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client[fastembed]>=1.17.1",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contextforge"]
